requirements.txt
absl-py==0.4.1
aniso8601==3.0.2
astor==0.7.1
click==6.7
Flask==1.0.2
Flask-Cors==3.0.6
Flask-RESTful==0.3.6
gast==0.2.0
grpcio==1.14.2
h5py==2.8.0
itsdangerous==0.24
Jinja2==2.10
Keras==2.1.6
Markdown==2.6.11
MarkupSafe==1.0
numpy==1.14.5
oyaml==0.6
pandas==0.23.4
protobuf==3.6.1
pyodbc==4.0.24
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
scikit-learn==0.19.2
scipy==1.1.0
setuptools==39.1.0
six==1.11.0
SQLAlchemy==1.2.11
tensorflow==1.4.0
termcolor==1.1.0
Werkzeug==0.14.1
xgboost==0.80
